mysite/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-debug-toolbar==6.0.0
sqlparse==0.5.3
